Implemented output to ResultTable and the use of the IJ progress bar.
Drift and shift of each image are shown at a ResultTable. The shift is
only added if an image shift is performed.
NormCrossCorrelation has been extended to show the progress of the
calculation by using IJ.showProgress(progress, progressSteps). A public
static method is used to set higher values of processSteps. this is
necessary because DriftDetectionPlugin can create more than one instance
of NormCrossCorrelation that should use the same process indicator.