Implemented stack creation and did some optimisations.
ExtendedStackToImage is a modified version of StackToImage. The extended
version gives the opportunity to sort the images.

A GenericDialog is used to select the mode of drift detection. A
YesNoCancelDialog was used until now.
Added a message that informs about the not yet implemented manual mode.
OptimisedStackShifter uses the right title if a new image is created.