Implemented automatic SR-EELS detection and the EFTEMjLogTool.
I used the existing SR-EELS correction and added it to EFTEMj. I have
done only a few changes. The next step is to add the manual detection.

EFTEMjLogTool is a small program to save a process log and show it at a
GenerigDialog. This is only a draft that needs to be extended.