Added scripts for Gnuplot and ImageJ (Fiji):
The Gnuplot scripts are used to process the results of
'SR-EELS_characterisation.ijm'. They create polynomials that are used to
describe the distortion.
'correctSR-EELS.ijm' and 'correctSR-EELS.py' perform the same task. They
use the polynomials and correct an SR-EELS dataset. The Python version
is much faster and it is able to correct the intensity.
correctSR-EELS will be replaced by a plugin written in Java.