Implemented the SR-EELS correction (no correction of intensity yet):
I compared the new results with the ones created with the ImageJ macro
and the Jython script.
The Java version is much faster than the Jython version, which is much
faster than the macro version.
The results for Java and Jython are identical at more than 97% of the
pixels. Pixels that differ are arranged along curves. The macro result
differs a lot, as it uses a slightly different correction method.

I use the GenericDialogPlus, that is part of the fiji-lib. I simply
copied the code, as I don't want my plugin to depend on Fiji.