Improvements to SR-EELS_Correction:
I removed ghost images that occured because of negative areas during
integration. 'SR-EELS_Exception.java' is part of this fix.
The correction can be performed on images with different binning values.
Yet only binning 4 was possible.
When reading the parameters file, lines with commends (char #) are
ignored.
The plugin estimates the remaining time.
The calculation of the bounding rectangle is more save now.