Improvements to SR-EELS_Correction:
I removed the oversampling and implemented a morphological filter. By
using dilate and erode holes in the binary image are closed. This is
much faster than oversampling and removes an intensity problem I was not
able to comprehend.
The display of the remaining time has a 5 seconds delay to avoid wrong
values at the start.