MC2015: Started to create figure 3 & figure 4
I used the plt files from my IMC2014 poster as a template.
Figure 3 shows only the borders of one SR-EELS dataset. It's nearly
finished.
Figure 4 is not finished yet.
There are some problems with code from Gnuplot 4.6 in Gnuplot 5.0.

The used measurements are the same as the one that can be downloaded.