SR-EELS_characterisation.js uses 2 methods to detect the spectrum:
First I implemented a new method that searches for the region that
exceeds a certain value.
The value has to decrease with increasing energy loss, as our camera has
a high blur and produces ghost images.

The second method uses thresholding and was already part of
SR-EELS_characterisation.ijm. There is still a bug, as the results are
different compared to the ijm version.