Fixing a bug in SR-EELS_Characterisation.js:
The results are nearly the same, as in the ijm version.
IJ.run(subImage.imp, "Bin...", "x=1 y=" + subImage.imp.getHeight() + "
bin=Average");
was causing the problems. In js I have to use the not binned image to
get correct results.
In ijm binning works fine. The reason is not known.

Adding the width to the result plot. I will not use 4 single plots for
each image, as the ijm version does.